Move upload function into userServices
